Move workflow script inside the package (backwards-compatible)
