+++ src/main/anovos/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+from .workflow import run
+
+run(config_path=sys.argv[1], run_type=sys.argv[2])
